+++ packages/api/lambdas/sf-event-sqs-to-db-records/write-granules.js
@@ -202,6 +202,7 @@ const writeGranuleAndFilesViaTransaction = async ({
   pdrCumulusId,
   fileUtils = FileUtils,
   trx,
+  updatedAt,
 }) => {
   const { files = [] } = granule;
   // This is necessary to set properties like
@@ -225,6 +226,7 @@ const writeGranuleAndFilesViaTransaction = async ({
     providerCumulusId,
     pdrCumulusId,
     processingTimeInfo,
+    updatedAt,
   });
 
   const upsertQueryResult = await upsertGranuleWithExecutionJoinRecord(
@@ -295,6 +297,7 @@ const writeGranule = async ({
   providerCumulusId,
   pdrCumulusId,
   granuleModel,
+  updatedAt = Date.now(),
 }) =>
   knex.transaction(async (trx) => {
     await writeGranuleAndFilesViaTransaction({
@@ -310,6 +313,7 @@ const writeGranule = async ({
       executionCumulusId,
       pdrCumulusId,
       trx,
+      updatedAt,
     });
     return granuleModel.storeGranuleFromCumulusMessage({
       granule,
@@ -322,6 +326,7 @@ const writeGranule = async ({
       workflowStatus,
       processingTimeInfo,
       queryFields,
+      updatedAt,
     });
   });